Fix problem instanciating queue connection.
Using Laravel Scout, it received this error with the original code:

Illuminate/Contracts/Container/BindingResolutionException with message 'Target [Illuminate/Database/ConnectionResolverInterface] is not instantiable while building [Mingalevme/Illuminate/UQueue/Connectors/DatabaseConnector].

I checked the code in the laravel framework and I modified it to work the same way.

It now works for me.